backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic-settings==2.6.1
python-docx==1.1.2
weasyprint==63.0
openai==1.57.0
anthropic==0.42.0
